Correct use of drop in apply-generic.
drop should only be called at the end of evaluation. This is needed
because we were initially calling apply-generic recursively to raise
the arguments.

Also, project and raise should not call into apply-generic after the
changes to drop the final answer because that would make no sense
(it reverses what we have just done).